requirements.txt
aniso8601==1.2.1
astroid==1.6.0
certifi==2017.11.5
chardet==3.0.4
click==6.7
flake8==3.5.0
Flask==0.12.2
flask-restplus==0.10.1
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
jsonschema==2.6.0
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.8.1
python-dateutil==2.6.1
python-dotenv==0.7.1
pytz==2017.2
requests==2.18.4
six==1.11.0
urllib3==1.22
virtualenv==15.1.0
Werkzeug==0.12.2
wrapt==1.10.11
